new code batch
fixed the navlink and where I select categories to where they accurately work

still working on getting the cart to work
