+++ COPY-vite-project-COPY5/.eslintrc.cjs
