+++ COPY-vite-project-COPY5/src/App.jsx
@@ -2,8 +2,8 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import SingleProduct from './Components/SingleProduct';
 import AllProducts from './Components/AllProducts';
 import Cart from './Components/CartPage/Cart';
-import Login from './Components/Login/Login';
-import Payment from './Components/Payment/Payment';
+import Login from './Components/LoginPage/Login';
+import Payment from './Components/PaymentPage/Payment';
 import NavBar from './Components/Navbar';
 
 function App() {
