+++ COPY-vite-project-COPY5/src/Components/LoginPage/Login.jsx
@@ -1,5 +1,5 @@
 // Login.jsx
-import React, { useState } from 'react';
+import { useState } from 'react';
 import './Login.css';
 
 export default function Login() {
