+++ COPY-vite-project-COPY5/src/Components/PaymentPage/Payment.jsx
@@ -1,5 +1,3 @@
-// Payment.jsx
-import React from 'react';
 import './Payment.css';
 
 export default function Payment() {
