+++ COPY-vite-project-COPY5/src/Components/ProductCard/ProductCard.jsx
@@ -1,5 +1,4 @@
 // ProductCard.jsx
-import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import "./ProductCard.css";
