+++ COPY-vite-project-COPY5/src/Components/SingleProduct.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { getProductbyId } from '../API';
 import ProductCard from './ProductCard/ProductCard';
